requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
constantly==15.1.0
cryptography==3.4.8
cycler==0.10.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.7
django-appconf==1.0.4
django-bulma==0.8.2
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-heroku==0.3.1
django-imagekit==4.0.2
django-recaptcha==2.0.6
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.2.0.32
opt-einsum==3.3.0
pilkit==2.0
Pillow==8.3.2
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
zope.interface==5.4.0
